requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==21.5b0
boto3==1.17.66
botocore==1.20.66
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flake8==3.9.1
identify==2.2.4
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.8.0
jmespath==0.10.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.4
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
pre-commit==2.12.1
py==1.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
s3transfer==0.4.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
zipp==3.4.0
